LLM Project Chatbot/app_backend.py
import pinecone
import pandas as pd
from io import StringIO
from langchain.embeddings.openai import OpenAIEmbeddings
from langchain.vectorstores import pinecone
from langchain_openai import ChatOpenAI
from langchain.chains import RetrievalQA
from langchain_experimental.tools.python.tool import PythonAstREPLTool
from langchain.agents import initialize_agent, Tool
from langchain.agents import AgentType
from langchain.chains import LLMMathChain
from settings import SETTINGS