LLM Project Chatbot/requirements.txt
langchain
streamlit
streamlit-chat
pinecone-client
pandas
python-dotenv